refactor:#76 modal 컴포넌트 위치 변경
